+++ src/Core/Dfe.Complete.Application/Notes/Interfaces/ITaskDataWriteRepository.cs
@@ -4,7 +4,7 @@ namespace Dfe.Complete.Application.Notes.Interfaces
 {
     public interface ITaskDataWriteRepository
     {
-        Task UpdateConversionAsync(ConversionTasksData conversionTasksData, CancellationToken cancellationToken);
-        Task UpdateTransferAsync(TransferTasksData transferTasksData, CancellationToken cancellationToken);
+        Task UpdateConversionAsync(ConversionTasksData conversionTasksData, DateTime dateTime, CancellationToken cancellationToken);
+        Task UpdateTransferAsync(TransferTasksData transferTasksData, DateTime dateTime, CancellationToken cancellationToken);
     }
 }